+++ build-types/components/collaborative-editing/use-yjs/formats/collab-caret/color-utils.d.ts
@@ -1,0 +1,7 @@
+/**
+ * Determine whether white text should be used on a given background color, based on YIQ contrast.
+ *
+ * @param {string} backgroundColorHex
+ */
+export function shouldUseWhiteText(backgroundColorHex: string): boolean;
+//# sourceMappingURL=color-utils.d.ts.map