+++ src/components/collaborative-editing/use-yjs/filters/block-selection/index.js
@@ -14,12 +14,16 @@ import './style.scss';
  */
 const addSelectionBorders = ( OriginalComponent ) => {
 	return ( props ) => {
-		const isSelected = useSelect(
+		const { isSelected, color } = useSelect(
 			( select ) => {
 				const peers = select( 'isolated/editor' ).getPeers();
-				return Object.values( peers ).some(
+				const matchedPeer = Object.values( peers ).find(
 					( peer ) => peer.start?.clientId === props.clientId && peer.end?.clientId === props.clientId
 				);
+				return {
+					isSelected: !! matchedPeer,
+					color: matchedPeer?.color,
+				};
 			},
 			[ props.clientId ]
 		);
@@ -27,11 +31,16 @@ const addSelectionBorders = ( OriginalComponent ) => {
 			<OriginalComponent
 				{ ...props }
 				className={ isSelected ? 'is-iso-editor-collab-peer-selected' : undefined }
+				wrapperProps={ {
+					style: {
+						'--iso-editor-collab-peer-block-color': color,
+					},
+				} }
 			/>
 		);
 	};
 };
 
 export const addFilterCollabBlockSelection = () => {
-	addFilter( 'editor.BlockListBlock', 'isolated-block-editor', addSelectionBorders );
+	addFilter( 'editor.BlockListBlock', 'isolated-block-editor', addSelectionBorders, 9 );
 };