+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Tweaks to collaborative editing UI.
+
 ## [2.4.1] - 2021-09-09
 
 ### Fixed